fix issues on feedback 16/06/2022

+++ Controllers/AkTerimaController.cs
@@ -479,7 +479,7 @@ namespace MSNK.Controllers
 
                 _cart.AddItem1(AkTerimaId,
                                result.JumLulus,
-                               result.AkCartaId);
+                               (int)result.AkCartaId);
 
                 return Json(new { result = "OK", record = result });
             }
